pyproject.toml
[project]
name = "intro-to-prefect"
description = "An introductory project to Prefect"
version = "0.1.0"
authors = [
    { name = "Shiza Azam", email = "shizaazam6@gmail.com" }
]
dependencies = [
    "prefect ~= 2.13.2",
    "httpx ~= 0.25.0",
    "psycopg2-binary ~= 2.9.7"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pre-commit"
]

[project.scripts]
pipeline = "pipeline.main:main"
